Implement basic popping

+++ core/src/ru/mofr/popballs/components/ClickComponent.java
@@ -2,7 +2,10 @@ package ru.mofr.popballs.components;
 
 import com.artemis.Component;
 
+/**
+ * x and y are in world space
+ */
 public class ClickComponent extends Component {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 }
